SAFRAN-434: Wring SQL generation on View's query change
Add and repair SQL generation tests.

Bug: SAFRAN-434

+++ reversers/database-synchroniser/plugins/org.obeonetwork.dsl.database.sqlgen.tests/input/07-column-type-changed/expectedSql/alter-tables.sql
@@ -1,7 +1,7 @@
 -- ==============================================================
 --  Table : TABLE1                                    
 -- ==============================================================
-ALTER TABLE TABLE1 MODIFY COL3 INTEGER NOT NULL;
+ALTER TABLE TABLE1 MODIFY COL3 LONG NOT NULL;
 
 ALTER TABLE TABLE1 MODIFY Col2 DATE NOT NULL;
